explore screen layout

+++ FoodicsRecipes/ReusableViews/RatingView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct RatingView: View {
-    var rating: Int
+    var rating: Double
     var reviewCount: Int
     
     var body: some View {
@@ -16,8 +16,8 @@ struct RatingView: View {
             Spacer()
             Text("\(String(format: "%.1f", rating))")
             ForEach(0..<5, id: \.self) { index in
-                Image(systemName: index < rating ? "star.fill" : "star")
-                    .foregroundColor(index < rating ? .yellow : .gray)
+                Image(systemName: index < Int(rating) ? "star.fill" : "star")
+                    .foregroundColor(index < Int(rating) ? .yellow : .gray)
             }
             Text("(\(reviewCount))")
                 .font(.subheadline)
